+++ src/lib/pages/account-details/components/tables/contracts/index.tsx
@@ -1,0 +1,2 @@
+export * from "./AdminContractTable";
+export * from "./InstantiatedContractTable";