+++ src/lib/pages/code-details/components/table/contracts/ContractTable.tsx
@@ -7,14 +7,13 @@ import { NoContracts } from "../NoContracts";
 import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
 import { TableContainer, TableHeader } from "lib/components/table";
+import { ContractTableRow } from "lib/components/table/contracts/ContractTableRow";
 import { useContractStore } from "lib/hooks";
 import {
-  useContractListByCodeId,
+  useContractListByCodeIdPagination,
   useContractListCountByCodeId,
-} from "lib/services/codeService";
+} from "lib/services/contractService";
 import type { ContractInfo, Option } from "lib/types";
-
-import { ContractTableRow } from "./ContractTableRow";
 
 interface ContractTableProps {
   codeId: number;
@@ -40,7 +39,7 @@ export const ContractTable = observer(({ codeId }: ContractTableProps) => {
     },
   });
 
-  const { data: rawCodeContracts } = useContractListByCodeId(
+  const { data: rawCodeContracts } = useContractListByCodeIdPagination(
     codeId,
     offset,
     pageSize